+++ lib/edgedb/connection/config.ex
@@ -289,7 +289,8 @@ defmodule EdgeDB.Connection.Config do
       transaction: from_config(:transaction),
       retry: from_config(:retry),
       connection: from_config(:connection),
-      pool: from_config(:pool)
+      pool: from_config(:pool),
+      state: from_config(:state)
     )
   end